% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_ancova.R
\name{bct.ancova}
\alias{bct.ancova}
\title{Batch Effect Correction with Analysis of Covariance \(ANCOVA\)}
\usage{
bct.ancova(peaks, info, minQC = 4, condition = c("", 0, 1, 2, "c"),
  method = c("lm", "rlm", "tobit"), QC = TRUE)
}
\arguments{
\item{peaks}{Input peak data table in pipeline format (samples are in rows, features
are in columns).}

\item{info}{Information table generated by bct.infogen.}

\item{minQC}{Minimum number of quality control samples per batch. Any batch containing
less than minQC quality control samples will be discarded for correction.}

\item{condition}{Missing data imputation option. Can be "" (none) to ignore NAs, 0 to
impute NAs with 0, 1 to impute NAs with half of the limit of detection, 2 to impute
NAs with the limit of detection, or "c" to automatically set a threshold for
censored regression. Note that if set to "c", the method argument will be set to
"tobit".}

\item{method}{Linear regression method option. Determines modelling approach applied
to compute the ANCOVA corrected peak intensity values. Can be "lm" to perform linear
model regression, "rlm" to perform robust linear regression, or "tobit" to perform
censored regression by tobit. Note that if set to "tobit", the contidion argument
will be set to "c".}

\item{QC}{Logical; should regression modelling be done with quality control samples
(QC == T) or with the study samples?}
}
\value{
Produces a numeric matrix containing the corrected peak intensities in the
  same format as the input.
}
\description{
Core element of the pipeline algorithm performing batch effect correction.
  Original implementation by R. Wehrens (Wehrens, R., Hageman, J. A., van Eeuwijk, F.,
  Kooke, R., Flood, P. J., Wijnker, E., … de Vos, R. C. H. (2016). Improved batch
  correction in untargeted MS-based metabolomics. Metabolomics, 12(5).
  https://doi.org/10.1007/s11306-016-1015-8).
  Uses an analysis of covariance framework containing three linear modelling approaches
  to estimate corrected values based on the applied regression.
}
