% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_bridge.R
\name{bct.bridge}
\alias{bct.bridge}
\title{Reformatting to original input format.}
\usage{
bct.bridge(peaks, info, backup, p.id, start.ord, gn = NULL, csv = F,
  n = NULL, wrap = T)
}
\arguments{
\item{peaks}{Peak data table (matrix) to reformat}

\item{info}{Information table generated by bct.infogen.}

\item{backup}{Backup object generated by bct.format.}

\item{p.id}{Column index of first sample data column in original input table.}

\item{start.ord}{Character vector containing the sample names in the order of the
original input table}

\item{gn}{Character vector containing the sample group names.}

\item{csv}{Logical; should a .csv file be saved?}

\item{n}{Filename prefix to add to the filename. Required if csv = T.}

\item{wrap}{Logical; should formatting to original input format be performed? If
wrap == F, function call only saves a .csv file with csv = T.}
}
\value{
Produces a character matrix containing the reformatted data table if wrap = T
  and saves a .csv file with the specified prefix. If wrap = F only saves a .csv file.
}
\description{
Reformats pipeline formatted objects to original input objects and saves
  .csv files.
}
