% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_eval_RW.R
\name{bct.eval_RW}
\alias{bct.eval_RW}
\title{Basic evaluation of batch effect correction}
\usage{
bct.eval_RW(peaks, info, arguments, plot.pca = F, plot.rep = F,
  plot.pdist = F, n = "")
}
\arguments{
\item{peaks}{Peak data to evaluate. Samples have to be in rows (as produced
by outputted by the BCT call)}

\item{info}{Information table on peaks, containing information on sample grouping,
injection sequence and batch labels}

\item{arguments}{Input arguments from BCT produced object.}

\item{plot.pca}{Logical; should a PCA score plot be produced?}

\item{plot.rep}{Logical; should a repeatability histogram be produced?}

\item{plot.pdist}{Logical; should a p-value histogram be produced?}

\item{n}{Plot title}
}
\value{
Returns a list object with the values specified in the arguments object.
}
\description{
Rudimentary evaluation function computing Bhattacharyya distance and
  plotting PCA the according score plots, computing repeatability between batches
  and plotting histograms, as well as featurewise p-values for sample groups,
  injection sequence and batch labels.
}
