% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_getseq.R
\name{bct.getseq}
\alias{bct.getseq}
\title{Extract injection sequence from sample names}
\usage{
bct.getseq(dat)
}
\arguments{
\item{dat}{Sample name vector}
}
\value{
Returns a numeric vector with the matched injection sequence.
}
\description{
Function that takes a vector containing sample names and extracts the
  injection sequence, returns a character vector.
}
