% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_in_args.R
\name{bct.in.args}
\alias{bct.in.args}
\title{Graphical user interface function to input function arguments}
\usage{
bct.in.args()
}
\value{
Assigns global variables used in bct.in.main.
}
\description{
Builds a graphical user interface depending on the choices taken in
  bct.in.methods. Takes all of the required function arguments used during pipeline
  execution.
}
