% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_in_main.R
\name{bct.in.main}
\alias{bct.in.main}
\title{Master graphical user interface input function}
\usage{
bct.in.main()
}
\value{
Returns a list object containing the user choices, intended for passing
  arguments within the pipeline.
}
\description{
Calls graphical user interface functions to ease interaction with
  the pipeline. First calls bct.in.methods, registering the correction and
  normalization methods as well as the choice of regression samples (QCs or
  study samples). Second calls bct.in.args to take all of the required inputs
  for correct pipeline execution.
}
