% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_in_methods.R
\name{bct.in.methods}
\alias{bct.in.methods}
\title{Graphical user interface funciton to input method selection}
\usage{
bct.in.methods()
}
\value{
Assigns global variables used in bct.in.main.
}
\description{
Initial graphical user interface to choose the batch effect correction
  and normalization methods, and the choice to perform batch correction at hand of
  quality control samples or study samples.
}
