% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_infogen.R
\name{bct.infogen}
\alias{bct.infogen}
\title{Generate information table}
\usage{
bct.infogen(data.in, bn, gn, p.id)
}
\arguments{
\item{data.in}{Peak data table within a list.}

\item{bn}{Character vector containing batch names as found within sample names.}

\item{gn}{Character vector containing group names as found within sample names
(determined in BCT call).}

\item{p.id}{Column index of first sample data column (usually named 'Raw abundance').}
}
\value{
Returns a list containing the origianl input along with an information table
  element.
}
\description{
Function to produce the required information table from original input
  table.
}
