% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_linearity.R
\name{bct.linearity}
\alias{bct.linearity}
\title{Feature plotting function}
\usage{
bct.linearity(data.in, select, QC = T, seq = T, I.g = 1, I.b = 2,
  I.s = 3, log.t = F, lgd = T, r.method = c("lm", "loess"),
  span = 0.75)
}
\arguments{
\item{data.in}{Input list containing two elements: the peak data table in pipeline
format and the corresponding information table generated by infogen.}

\item{select}{Feature indices to plot. Can be a vector with multiple indices.}

\item{QC}{Logical; plot regression lines at hand of quality contol samples (QC = T)
or at hand of study samples?}

\item{seq}{Logical; plot along injection sequence indices (seq = T) or along batch
labels?}

\item{I.g}{Information table index of grouping column}

\item{I.b}{Information table index of batch label column}

\item{I.s}{Information table index of injection sequence column}

\item{log.t}{Logical; is the input peak data log-transformed?}

\item{lgd}{Logical; should a legend be included in the plot?}

\item{r.method}{Regression method to plot trendline. Can be "lm" for linear model or
"loess" for locally weighted scatterplot smoothing.}

\item{span}{Span parameter if r.method = "loess". Decides the degree of smoothing.}
}
\value{
Produces a ggplot object.
}
\description{
Plotting function that takes bct formatted LC-MS data and plots selected
  feature intensities against either injection sequence or batch labels. Uses the
  ggplot package for plotting.
}
