% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_loess_LA.R
\name{bct.loess}
\alias{bct.loess}
\title{Univariate locally weighted scatterplot smoothing (LOESS).}
\usage{
bct.loess(data.in, info, bn, LOESS_SPAN = 0.75, LOESS_WARNING = F,
  normplot = F, pn = NULL)
}
\arguments{
\item{data.in}{Originally the output of InputAndProcessCSV. Within the pipeline,
data.in is passed from BCT_brige.R, containing the data as prepared by the
preceding functions.}

\item{info}{Information table generated by bct.infogen.}

\item{bn}{Batch labels as found in the sample names.}

\item{LOESS_SPAN}{Distance parameter defining the window of consideration for
weighting of the response variable. Default = 0.75}

\item{LOESS_WARNING}{Logical; should a warning be produced by the call, if any occur?}

\item{normplot}{Logical; should normalization plots for each feature be produced and
saved as .png?}

\item{pn}{Plot file name suffix \\(for proper batch naming\\)}
}
\value{
Returns a numeric matrix containing the loess corrected peak intensity values.
}
\description{
Applies loess correction for intra-batch removal of technical variability.
  Original implementation by Stefan Zahnd for the LaunchAnalyzer toolkit.
}
