% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_mod.R
\name{bct.mod}
\alias{bct.mod}
\title{Modification of peak data}
\usage{
bct.mod(data.in, mod = c("", 1, 2, 3), log = c("", "e", 2, 10), R = F)
}
\arguments{
\item{data.in}{Peak data table formatted by bct.format.}

\item{mod}{Modification option. Modification means addition of one or multiple values. Can be "" for no modification, 1 for addition of limit of detection > 0, 2 for addition of the data mean, or 3 for the feature-wise addition of feature means.}

\item{log}{Logarithm transformation option. Can be "" for no log transformation, "e"
for ln transformation, 2 for log2 transformation, or 10 for log10 transformation.}

\item{R}{Logical; should modifications be applied (R == F) or reverted?}
}
\value{
Returns a numeric matrix with modified peak intensity values.
}
\description{
Applies modification through addition of values as well as logarithm
  transformation of peak intensity data. To avoid negative infinity values (log of 0),
  the addition is always performed before log transformation. When reverting, the
  logarithm is reverted before subtracting the modification value.
}
