% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_prep.R
\name{bct.prep}
\alias{bct.prep}
\title{Data preparation}
\usage{
bct.prep(data.in)
}
\arguments{
\item{data.in}{Input data specification of table to prepare. Can be a list containing
peak data and information table (bct.infogen), or a character object containing the
filename of the input data.}
}
\value{
Returns a list object containing either two or one elements, depending on the
  input object length.
}
\description{
Prepares input data (specified in either list with table objects or
  filename) for passing to bct.format. Reads .csv files if input is a filename.
}
