% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_pval.R
\name{bct.pval}
\alias{bct.pval}
\title{Compute featurewise p-values of LC-MS data by analysis of variance (ANOVA).}
\usage{
bct.pval(peaks, info, PLOT = TRUE, PDF = FALSE, csv = FALSE, t = "",
  adj = F, plotn = c("Group", "Batch", "SeqNr", "All"))
}
\arguments{
\item{peaks}{Input peak data table (matrix)}

\item{info}{Information table generated by bct.infogen.}

\item{PLOT}{Logical; should p-value histograms be produced?}

\item{PDF}{Logical; should a pdf with p-value histograms be produced. Obsolete
command within pipeline, set to F.}

\item{csv}{Logical; should a .csv file containing the featurewise p-values be produced?}

\item{t}{Plot title}

\item{adj}{Logical; should the computed p-values be adjusted (Bonferroni-Holm)?}

\item{plotn}{Plotting option to compute p-values by information present in information
table. Can be "Group" to compute p-values between sample groups, "Batch" for
p-values between batches, "SeqNr" for p-values between injection sequence indices,
or "All" to compute all of the mentioned p-values.}
}
\value{
Generates a table object containing p-values of all features.
}
\description{
Performs ANOVA on the peak data using the information table to compute
  p-values per feature. Used to univariately assess the presence of batch effects, as
  well as to control if differences between groups (actual variation of interest) is
  perserved (courtesy of Prof. C. T. Nakas, Laboratory of Biometry, University of
  Thessaly, Volos, Greece).
}
