% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_splitbind2.R
\name{bct.splitbind2}
\alias{bct.splitbind2}
\title{Split or Combine LC-MS Datasets}
\usage{
bct.splitbind2(data.in, info, p.i, bn, start.ord, bind, form, p.id = NULL,
  gn = NULL)
}
\arguments{
\item{data.in}{Input data to split or bind. Has to be a list containing the peak table
to split or the peak tables to bind.}

\item{info}{Information table to the peak data intended to split or bind.}

\item{p.i}{Index or name of peak table within input list}

\item{bn}{Batch naming as found in information table}

\item{bind}{Logical; should the data be split (bind = F) or bound together (bind = T)}

\item{form}{Logical; has the data been formatted by bct.format (form = T) before
inputting to bct.splitbind2?}

\item{p.id}{Required if form = F; column index of first data column (usually named
'Raw abundance').}

\item{gn}{Required if form = F; sample groups as found in information table, used
to correctly label the samples by grouping.}
}
\value{
Produces a list containing data table elements split by batch labels if
  bind = F. Produces a table object with combined tables if bind = T.
}
\description{
Function in order to prepare peak data tables for intra-batch
  normalization, as well as returning them to their original composition.
}
