% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_tabwrap.R
\name{bct.tabwrap}
\alias{bct.tabwrap}
\title{Simplified class conversion funciton}
\usage{
bct.tabwrap(tab.in, type.out)
}
\arguments{
\item{tab.in}{Input table object to be converted.}

\item{type.out}{Output class of table object. Is a character vector containing
the table element class as the first and and the table class as second element
(i.e. c("character", "data.frame")). First element can be "numeric", "character",
or "factor". Second element can be "matrix", "data.frame", or "data.table". Uses
the data.table package.}
}
\value{
Returns the input table with specified classes.
}
\description{
Function mainly aiming at simple class conversion of table objects passed
  within the pipeline.
}
