% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enterophage.R
\docType{data}
\name{enterophage}
\alias{enterophage}
\title{Enterobacteria_phage_lambda}
\format{
An object of class \code{"character"}.
}
\usage{
enterophage
}
\description{
This dataset contains the 48502 base-pair-long genome of the bacteriophage lambda virus.
The bacteriophage lambda is a parasite of the intestinal bacterium Escherichia coli.
This virus is a benchmark sequence for the comparison of segmentation algorithms.
}
\references{
{Enterobacteria phage lambda, complete genome.}
(\href{https://www.ncbi.nlm.nih.gov/nuccore/9626243}{Enterobacteria_phage_lambda})
}
\keyword{datasets}
