% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pewee.R
\docType{data}
\name{pewee}
\alias{pewee}
\title{Pewee birdsong}
\format{
An object of class \code{"character"}.
}
\usage{
pewee
}
\description{
The twilight song of the wood pewee bird can be described as a sequence consisting of an arrangement of musical phrases taken 
from an alphabet of three specific, distinct phrases.
The dataset consists of a single continuous song by a wood pewee, 
of length n = 1327 phrases.
}
\examples{
BCT(pewee, 5)
}
\references{
{W. Craig. The song of the wood pewee (Myiochanes virens Linnaeus): A study of bird music. New York State Museum Bulletin No. 334. University of the State of New York, Albany, NY, 1943.}
(\href{https://www.worldcat.org/title/song-of-the-wood-pewee-myiochanes-virens-linnaeus-a-study-of-bird-music/oclc/2133859}{craig})
}
\keyword{datasets}
