% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_bct.R
\name{show_tree}
\alias{show_tree}
\title{Plot tree with given contexts}
\usage{
show_tree(s)
}
\arguments{
\item{s}{vector containing the contexts of the leaves of the desired tree.}
}
\value{
plot of the desired tree model.
}
\description{
Plots a tree depicting a model with the given set of contexts.
}
\examples{
# Construct an example vector:
r <- c("a", "ab", "aab", "b", "ba")

show_tree(r)

# If the input contains digits:
q <- c(11,1,0)

show_tree(q)
}
\seealso{
\code{\link{BCT}}, \code{\link{draw_models}}
}
