% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simian_40.R
\docType{data}
\name{simian_40}
\alias{simian_40}
\title{simian_40}
\format{
An object of class \code{"character"}.
}
\usage{
simian_40
}
\description{
This dataset contains the 5243 base-pair-long genome of the simian vacuolating virus 40 (SV40).
SV40 is a polyomavirus that is found in both monkeys and humans. The expression of SV40 genes is regulated by two major transcripts (early and late), suggesting the presence of a single major change-point in the entire genome.
}
\references{
{}
(\href{https://www.ncbi.nlm.nih.gov/nuccore/J02400}{SV40})
}
\keyword{datasets}
