% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three_changes.R
\docType{data}
\name{three_changes}
\alias{three_changes}
\title{three_changes}
\format{
An object of class \code{"character"}.
}
\usage{
three_changes
}
\description{
This dataset contains synthetic-generated data. The sequence has three change-points located at 2500, 3500 and 4000.
The tree models and associated parameters are chosen to be quite similar, so that the segmentation problem is nontrivial.
}
\references{
{V. Lungu, I. Papageorgiou and I. Kontoyiannis. Change-point Detection and Segmentation of Discrete Data using Bayesian Context Trees. arxiv.2203.04341}
(\href{https://arxiv.org/abs/2203.04341}{Segmentation})
}
\keyword{datasets}
