% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bct_master.R
\name{BCT}
\alias{BCT}
\title{Master function of the batch correction tool}
\usage{
BCT()
}
\value{
The output object is a list containing three lists: DATA, containing all of
  the data passed within the pipeline; DATA.FORMATTED, containing the tables that
  are saved as .csv files; and ARGS, containing the input arguments as well as
  arguments added during the execution.
}
\description{
Main function wrapper containing the intended workflow of the batch
  effect correction algorithm. Calling BCT will initiate the input interface
  requiring user interaction. The chosen inputs are then passed within the pipeline
  to compute corrections as well as perform the required modifications and formatting.
  The final results are then consistently outputted to separate directories created
  during the process.
}
